feat: 사용자 인증이 필요한 api에 jwt_required 추가.

+++ repository/post_repository.py
@@ -5,17 +5,23 @@ def read_post_detail(id):
     return Post.objects.get_or_404(id=id)
 
 
-def create(title, content, date, current_user):
-    post = Post(title=title, content=content, create_date=date, user=current_user).save()
+def create(title, content, date, current_user_id):
+    post = Post(title=title, content=content, create_date=date, user=current_user_id).save()
     id = post.id
     return id
 
 
-def modify(id, title, content, modify_date, current_user):
+def modify(id, title, content, modify_date, current_user_id):
     post = Post.objects.get_or_404(id=id)
-    post.update(title=title, content=content, modify_date=modify_date, user=current_user)
+    if post.user == current_user_id:
+        post.update(title=title, content=content, modify_date=modify_date)
+        return True
+    return False
 
 
-def delete(id, current_user):
+def delete(id, current_user_id):
     post = Post.objects.get_or_404(id=id)
-    post.delete()
+    if post.user == current_user_id:
+        post.delete()
+        return True
+    return False
